Convert majority of size_t to uint64_t for 32-bit compatibility
We need to get puffin to work correctly on 32-bit machines. This patch converts
appropriate size_t types to uint64_t to make sure when we are working with large
files in 32bit machines, nothing breaks.

The 32 bit tests actually found some errors. For example BitReader and BitWriter
should have their size and indices kept in 64 bit integers otherwise a multiply
by 8 (to figure out the bit-based length) will overflow. Although it has not
been tested for puffdiff and puffpatch.

Bug: 73781483
Test: unittests pass
Test: puffhuff on a >5G zlib file passes
Test: build with 32-bit compiler, and puffhuff on >1GB passes

+++ src/bit_reader.cc
@@ -64,7 +64,7 @@ size_t BufferBitReader::Offset() const {
   return index_ - in_cache_bits_ / 8;
 }
 
-size_t BufferBitReader::OffsetInBits() const {
+uint64_t BufferBitReader::OffsetInBits() const {
   return (index_ * 8) - in_cache_bits_;
 }
